refactor(core): move rate limit configuration to RateLimitConfig class
